requirements.txt
streamlit>=1.31.0
google-genai>=0.1.0
pydantic>=2.5.0
Pillow>=10.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
numpy>=1.24.0
